Change address entity to contact, create contactcontroller and move some action from membercontroller to contactcontroller; edit, delete, change default contact.

+++ WebApp/Models/Contact.cs
@@ -7,7 +7,7 @@ namespace WebApp.Models
 {
     public class Contact
     {
-        public int ContactId { get; set; }
+        public short ContactId { get; set; }
         public string AddressHome { get; set; }
         public short ProvinceId { get; set; }
         public string ProvinceName { get; set; }
